Work on units to include
Started working on what units should be included for structural analysis. The ifc standard includes an IfcThingMeasure for anything that could possible be measured even if it has the same dimensionality as something else. I'm not sure why they do it, but I will assume just use the most specific measure available.